hopr_transport_session::balancer

Module rate_limiting

Structs

RateController

Controller for RateLimitedStream to allow dynamic controlling of the stream’s rate.

RateLimitedSink

A sink adapter that allows ingesting items at a controlled rate, with dynamic rate adjustment.

RateLimitedStream

A stream adapter that yields elements at a controlled rate, with dynamic rate adjustment.

Enums

SinkState 🔒

StreamState 🔒

Traits

RateLimitSinkExt

Extension trait to add rate limiting to any sink.

RateLimitStreamExt

Extension trait to add rate limiting to any stream

Functions

rate_from_delay 🔒