hopr_transport_session

Module initiation

This module defines the Start sub-protocol used for HOPR Session initiation and management.

Structs

StartErrorType

Error message in the Start protocol.

StartEstablished

Message of the Start protocol that confirms the establishment of a session.

StartInitiation

The session initiation message of the Start protocol.

Enums

StartErrorReason

Lists all Start protocol error reasons.

StartProtocol

Lists all messages of the Start protocol for a session establishment with T as session identifier.

Type Aliases

StartChallenge

Challenge that identifies a Start initiation protocol message.