hopr_transport_session::utils

Function insert_into_next_slot

pub(crate) fn insert_into_next_slot<F, K, U, V>(
    cache: &Cache<K, V>,
    generator: F,
    value_fn: U,
    max_capacity: Option<u64>,
) -> Option<(K, V)>
where
    F: FnMut(Option<K>) -> K,
    K: Copy + Hash + Eq + Send + Sync + 'static,
    U: FnOnce(K) -> V,
    V: Clone + Send + Sync + 'static,

This function will use the given generator to generate an initial seeding key. It will check whether the given cache already contains a value for that key, and if not, calls the generator (with the previous value) to generate a new seeding key and retry. The function either finds a suitable free slot, inserting value generated by value_fn and returns the found key, or terminates with None when gen returns the initial seed again.