Crate hopr_transport_session

Session object providing the session functionality over the HOPR transport

The session proxies the user interactions with the transport in order to hide the advanced interactions and functionality.

Re-exports

pub use types::Session;

pub use types::SessionId;

pub use types::SESSION_USABLE_MTU_SIZE;

Modules

errors

traits

types

Structs

SessionClientConfig

Configuration for the session.

Enums

Capability

PathOptions

Send options for the session.