Crate hopr_transport_session

Session object providing the session functionality over the HOPR transport

The session proxies the user interactions with the transport in order to hide the advanced interactions and functionality.

Modules

errors

traits

Structs

CapabilityIter

An iterator over the variants of Capability

IncomingSession

Wrapper for incoming Session along with other information extracted from the Start protocol during the session establishment.

Session

SessionClientConfig

Configuration for the session.

SessionId

Unique ID of a specific session.

SessionManager

Manages lifecycles of Sessions.

SessionManagerConfig

Configuration for the SessionManager.

Enums

Capability

Capabilities of a session.

DestinationRouting

Routing information containing forward or return routing options.

DispatchResult

Indicates the result of processing a message.

IpOrHost

Implements a host name with port. This could be either a DNS name with port or an IP address with port represented by std::net::SocketAddr.

IpProtocol

Lists some of the IP protocols.

ResolvedTransportRouting

Contains the resolved routing information for the packet.

RoutingOptions

Represents routing options in a mixnet with a maximum number of hops.

SealedHost

Contains optionally encrypted IpOrHost.

SessionTarget

Defines what should happen with the data at the recipient where the data from the established session are supposed to be forwarded to some target.

Constants

SESSION_USABLE_MTU_SIZE

Inner MTU size of what the HOPR payload can take (payload - peer address - application_tag)

Functions

transfer_session

Convenience function to copy data in both directions between a Session and arbitrary async IO stream. This function is only available with Tokio and will panic with other runtimes.

unwrap_chain_address

wrap_with_chain_address

Type Aliases

ServiceId

Describes a node service target. These are specialized SessionTargets that are local to the Exit node and have different purposes, such as Cover Traffic.