Crate hopr_transport_session

Session object providing the session functionality over the HOPR transport

The session proxies the user interactions with the transport in order to hide the advanced interactions and functionality.

Re-exports

pub use types::IncomingSession;

pub use types::Session;

pub use types::SessionId;

pub use types::SESSION_USABLE_MTU_SIZE;

Modules

errors

initiation

This module defines the Start sub-protocol used for HOPR Session initiation and management.

traits

types

Structs

CapabilityIter

An iterator over the variants of Capability

SessionClientConfig

Configuration for the session.

SessionManager

Manages lifecycles of Sessions.

SessionManagerConfig

Configuration for the SessionManager.

Enums

Capability

Capabilities of a session.

DispatchResult

Indicates the result of processing a message.

IpOrHost

Implements a host name with port. This could be either a DNS name with port or an IP address with port represented by std::net::SocketAddr.

IpProtocol

Lists some of the IP protocols.

RoutingOptions

Represents routing options in a mixnet with a maximum number of hops.

SealedHost

Contains optionally encrypted IpOrHost.