hopr_transport_session

Module types

Structs

ByteCapabilities

Wrapper for Capabilities that makes conversion to/from u8 possible.

HoprSession

Represents the Session protocol socket over HOPR.

HoprSessionConfig

Configures the Session protocol socket over HOPR.

IncomingSession

Wrapper for incoming HoprSession along with other information extracted from the Start protocol during the session establishment.

Enums

ClosureReason

Indicates the closure reason of a HoprSession.

SessionTarget

Defines what should happen with the data at the recipient where the data from the established session are supposed to be forwarded to some target.

Constants

SESSION_APPLICATION_TAG

Constant application tag used for all sessions. Previously tags were dynamically allocated per session.

SESSION_SOCKET_CAPACITY 🔒

Traits

AsyncReadWrite 🔒

Helper trait to allow Box aliasing

Functions

caps_to_ack_mode 🔒

Type Aliases

HoprStartProtocol

Start protocol instantiation for HOPR.

ServiceId

Describes a node service target. These are specialized SessionTargets that are local to the Exit node and have different purposes, such as Cover Traffic.

SessionId

Unique ID of a specific Session.