hopr_transport_session

Module types

Structs

IncomingSession

Wrapper for incoming Session along with other information extracted from the Start protocol during the session establishment.

InnerSession

Session

SessionId

Unique ID of a specific session.

Enums

SessionTarget

Defines what should happen with the data at the recipient where the data from the established session are supposed to be forwarded to some target.

Constants

SESSION_USABLE_MTU_SIZE

Inner MTU size of what the HOPR payload can take (payload - peer_id - application_tag)

Functions

transfer_session

Convenience function to copy data in both directions between a Session and arbitrary async IO stream. This function is only available with Tokio and will panic with other runtimes.

unwrap_offchain_key

wrap_with_offchain_key

Type Aliases

ServiceId

Describes a node service target. These are specialized SessionTargets that are local to the Exit node and have different purposes, such as Cover Traffic.